+++ test/integration/LdpHandlerWithoutAuth.test.ts
@@ -2,7 +2,7 @@ import { createReadStream } from 'fs';
 import fetch from 'cross-fetch';
 import type { Quad } from 'n3';
 import { DataFactory, Parser, Store } from 'n3';
-import { joinFilePath, PIM, RDF } from '../../src/';
+import { joinFilePath, joinUrl, PIM, RDF } from '../../src/';
 import type { App } from '../../src/';
 import { LDP } from '../../src/util/Vocabularies';
 import {
@@ -725,5 +725,31 @@ describe.each(stores)('An LDP handler allowing all requests %s', (name, { storeC
 
     // DELETE
     await deleteResource(resourceUrl);
+  });
+
+  it('supports range requests.', async(): Promise<void> => {
+    const resourceUrl = joinUrl(baseUrl, 'range');
+    await putResource(resourceUrl, { contentType: 'text/plain', body: '0123456789' });
+
+    let response = await fetch(resourceUrl, { headers: { range: 'bytes=0-5' }});
+    expect(response.status).toBe(206);
+    expect(response.headers.get('content-range')).toBe('bytes 0-5/10');
+    expect(response.headers.get('content-length')).toBe('6');
+    await expect(response.text()).resolves.toBe('012345');
+
+    response = await fetch(resourceUrl, { headers: { range: 'bytes=5-' }});
+    expect(response.status).toBe(206);
+    expect(response.headers.get('content-range')).toBe('bytes 5-9/10');
+    expect(response.headers.get('content-length')).toBe('5');
+    await expect(response.text()).resolves.toBe('56789');
+
+    response = await fetch(resourceUrl, { headers: { range: 'bytes=-4' }});
+    expect(response.status).toBe(206);
+    expect(response.headers.get('content-range')).toBe('bytes 6-9/10');
+    expect(response.headers.get('content-length')).toBe('4');
+    await expect(response.text()).resolves.toBe('6789');
+
+    response = await fetch(resourceUrl, { headers: { range: 'bytes=5-15' }});
+    expect(response.status).toBe(416);
   });
 });