+++ test/unit/http/ldp/GetOperationHandler.test.ts
@@ -10,7 +10,7 @@ import type { ResourceStore } from '../../../../src/storage/ResourceStore';
 import { NotImplementedHttpError } from '../../../../src/util/errors/NotImplementedHttpError';
 import { NotModifiedHttpError } from '../../../../src/util/errors/NotModifiedHttpError';
 import { updateModifiedDate } from '../../../../src/util/ResourceUtil';
-import { CONTENT_TYPE, HH } from '../../../../src/util/Vocabularies';
+import { CONTENT_TYPE, HH, SOLID_HTTP } from '../../../../src/util/Vocabularies';
 
 describe('A GetOperationHandler', (): void => {
   let operation: Operation;
@@ -58,6 +58,18 @@ describe('A GetOperationHandler', (): void => {
     expect(result.statusCode).toBe(200);
     expect(result.metadata).toBe(metadata);
     expect(metadata.get(HH.terms.etag)?.value).toBe('ETag');
+    expect(result.data).toBe(data);
+    expect(store.getRepresentation).toHaveBeenCalledTimes(1);
+    expect(store.getRepresentation).toHaveBeenLastCalledWith(operation.target, preferences, conditions);
+  });
+
+  it('returns 206 if the result is a partial stream.', async(): Promise<void> => {
+    metadata.set(SOLID_HTTP.terms.unit, 'bytes');
+    metadata.set(SOLID_HTTP.terms.start, '5');
+    metadata.set(SOLID_HTTP.terms.end, '7');
+    const result = await handler.handle({ operation });
+    expect(result.statusCode).toBe(206);
+    expect(result.metadata).toBe(metadata);
     expect(result.data).toBe(data);
     expect(store.getRepresentation).toHaveBeenCalledTimes(1);
     expect(store.getRepresentation).toHaveBeenLastCalledWith(operation.target, preferences, conditions);